chore: slice workflow utils
